no code warnings

+++ VerbatmIOS/Verbatm/FeedTableViewController.m
@@ -84,7 +84,7 @@
 }
 
 - (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section {
-    NSLog(@"self.FollowingProfileList.count = %d", self.FollowingProfileList.count);
+    NSLog(@"self.FollowingProfileList.count = %lu", (unsigned long)self.FollowingProfileList.count);
     return self.FollowingProfileList.count;
 }
